+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased] - TBD
 ### Added
 - Added routine `Its_Time_For_Diag` to `GeosUtil/time_mod.F90`
+- Added error trap to prevent GC-Classic History collections from being updated more frequently than the diagnostic timestep
 
 ### Changed
 - Updated `Interfaces/GCClassic/main.F90` so that diagnostic archival is done once per diagnostic timestep (i.e. the larger of the chemistry timestep or dynamic timestep)